+++ seqra-java-semgrep/samples/src/main/java/issues/i95/SpecificData.java
@@ -1,0 +1,7 @@
+package issues.i95;
+
+public class SpecificData {
+    public String getData() {
+        return "I could be controlled from the outside!";
+    }
+}