+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2025.10.02
+### feat: Add taint labels, tests, and optimize automata
+- Initial support for taint labels
+- Add test for `label` and `requires` keywords
+- Fix null pointer bug for first line retrieval
+- Remove redundant automata edges
 ## 2025.10.01
 ### feat: Publish release
 - feat: Publish release